@@ -15,16 +15,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "4.2.6",
-    "@angular/common": "4.2.6",
-    "@angular/compiler": "4.2.6",
-    "@angular/core": "4.2.6",
-    "@angular/forms": "4.2.6",
-    "@angular/http": "4.2.6",
-    "@angular/platform-browser": "4.2.6",
-    "@angular/platform-browser-dynamic": "4.2.6",
-    "@angular/router": "4.2.6",
-    "@ng-bootstrap/ng-bootstrap": "1.0.0-alpha.27",
+    "@angular/animations": "4.3.1",
+    "@angular/common": "4.3.1",
+    "@angular/compiler": "4.3.1",
+    "@angular/core": "4.3.1",
+    "@angular/forms": "4.3.1",
+    "@angular/http": "4.3.1",
+    "@angular/platform-browser": "4.3.1",
+    "@angular/platform-browser-dynamic": "4.3.1",
+    "@angular/router": "4.3.1",
+    "@ng-bootstrap/ng-bootstrap": "1.0.0-alpha.28",
     "angular-split": "^0.2.0",
     "animate.css": "^3.5.2",
     "bootstrap": "4.0.0-alpha.6",
@@ -44,9 +44,9 @@
     "zone.js": "^0.8.4"
   },
   "devDependencies": {
-    "@angular/cli": "1.2.0",
-    "@angular/compiler-cli": "4.2.6",
-    "@angular/language-service": "4.2.6",
+    "@angular/cli": "1.2.1",
+    "@angular/compiler-cli": "4.3.1",
+    "@angular/language-service": "4.3.1",
     "@types/codemirror": "^0.0.38",
     "@types/dagre": "^0.7.34",
     "@types/jasmine": "2.5.45",
